pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rs3-study-curvature"
version = "0.1.0"
description = "Curvature study for RS3: BD TOPO + DEM → v^2/r"
authors = [{ name = "Sébastien EDET" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "geopandas>=0.14",
  "shapely>=2.0",
  "pyproj>=3.6",
  "rasterio>=1.3",
  "numpy>=1.26",
  "pandas>=2.1",
  "matplotlib>=3.8",
  "scipy>=1.11",
  "pytest>=7.4",
  "PyYAML>=6.0",
  "tqdm>=4.66",
]

[project.optional-dependencies]
docs = [
  "mkdocs>=1.6",
  "mkdocs-material>=9.5",
  "mkdocstrings[python]>=0.25",
  "tqdm>=4.66",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]

[project.scripts]
rs3c = "rs3_study_curvature.cli.main:app"

# pyproject.toml
[tool.ruff]
line-length = 150
extend-exclude = ["out", "data", "scripts"]

[tool.ruff.lint]
select = ["E", "F"]
ignore = []
